Update tsconfig.json
typescript-config/consistent-casing checks if the property forceConsistentCasingInFileNames is enabled in your TypeScript configuration file (i.e. tsconfig.json).

Why is this important?
If you are working on a project where developers use different OSes (e.g.: Windows and Linux), enabling this property will prevent issues related to the file casing and how the files are referenced in the code.